Graph implementation style fixes, corrections, and refactored tests
